debug fest pt1

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: doberste <doberste@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 10:49:53 by doberste          #+#    #+#             */
-/*   Updated: 2025/10/02 13:49:52 by doberste         ###   ########.fr       */
+/*   Updated: 2025/10/08 13:13:07 by doberste         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *str, int search_str, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
